add --hap-cluster: best-comp (new default), garud-shuffle, soft-em
When a haplotype carries missing genotypes it can be compatible with several
distinct haplotypes -- missing sites act as wildcards -- and the rule that
picks one of them sets the haplotype frequency spectrum. That rule was
hardcoded, and it was the shuffle-and-absorb variant whose own comment says it
is for testing order-dependence. On one window of the test data with 5% of
genotypes missing, 120 runs differing only in --seed gave H12 between 0.066 and
0.100 at --match-tol 0 and between 0.132 and 0.390 at --match-tol 1.

  --hap-cluster best-comp      (default) each haplotype joins the most frequent
                               class it is compatible with, best-observed
                               haplotypes first; deterministic, no seed
                garud-shuffle  the previous behaviour, for reproducing old runs
                soft-em        EXPERIMENTAL; splits an ambiguous haplotype's
                               count across compatible classes in proportion to
                               their frequencies, iterated to convergence

Both new rules were written against Python implementations that had been
validated against this program (garud-shuffle: identical outcome support and
KS p 0.41-1.00 on class count and H12 over 120 seeds vs 400 orderings), and
the C++ reproduces those implementations to the spectra file's print precision
on four window/tolerance combinations.

HaplotypeFrequencySpectrum::sortedCount and ::size are now double, since
soft-em divides a count across classes. Every statistic already cast to double
before dividing, so integral counts give bit-identical results -- the whole
suite passed unchanged at that step.

Chosen on evidence, not preference. Simulation (msprime, 117-SNP windows, 200
haplotypes, 120 neutral + 120 sweep windows, genotypes masked at random per
individual per site) scored each rule against the true spectrum of the same
window: mean L1 on the K=10 spectrum at 1/2/5/10% missing is 0.0046/0.0077/
0.0198/0.0379 for garud-shuffle, 0.0025/0.0045/0.0122/0.0239 for best-comp and
0.0036/0.0058/0.0123/0.0193 for soft-em. Connected components over the
compatibility graph was also tested and rejected: it inflates H12 in sweep
windows by 0.215 on a true 0.261, the direction that manufactures sweep
signals, so it is not offered as an option.

best-comp is also faster than the rule it replaces -- it compares each
haplotype against the existing classes rather than against every other
haplotype -- 11.4 s against 107.8 s on chr22 of the example with 2% of
genotypes blanked, one thread.

Without missing genotypes and at --match-tol 0 all three are identical
(md5-identical spectra on the chr22 example), so complete-data runs are
unaffected. The two missing-data goldens are pinned to garud-shuffle to keep
that path covered; new cases cover best-comp (including that it is the default
and that --seed does not change it), soft-em, and that all three rules
conserve the haplotype count. make check 23/23.

+++ src/lassip-data.h
@@ -185,9 +185,11 @@ struct PopData
 
 struct HaplotypeFrequencySpectrum {
   map<string,double> hap2count;
-  //multimap<int,string> count2hap;
-  int *sortedCount;
-  int size;
+  //Class sizes, largest first. Double rather than int because --hap-cluster
+  //soft-em divides an ambiguous haplotype's count across the classes it could
+  //belong to; every other path puts whole numbers here.
+  double *sortedCount;
+  double size;
   int numClasses;
 };
